changed datatype of amount from int to float.
changed time in log table from 12 hour format to 24 hour format.
removed timestamp from wallets and categories tables.
Added activity and intent for report.

+++ app/src/main/java/com/example/myapplication/WalletViewActivity.java
@@ -71,7 +71,7 @@ public class WalletViewActivity extends AppCompatActivity {
             TableRow tr = new TableRow(this);
             tr.setLayoutParams(getLayoutParams());
             tr.addView(getTextView(wallet.getId(), wallet.getName(), Color.WHITE, Typeface.NORMAL, ContextCompat.getColor(this, R.color.colorAccent)));
-            tr.addView(getTextView(wallet.getId() + walletList.size(), Integer.toString(wallet.getAmount()), Color.WHITE, Typeface.NORMAL, ContextCompat.getColor(this, R.color.colorAccent)));
+            tr.addView(getTextView(wallet.getId() + walletList.size(), Float.toString(wallet.getAmount()), Color.WHITE, Typeface.NORMAL, ContextCompat.getColor(this, R.color.colorAccent)));
             wallets.addView(tr, getTblLayoutParams());
         }
